google-maps-navigation-benchmark/requirements.txt
crewai>=0.74.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20
litellm>=1.0.0
ollama>=0.1.0

